requirements.txt
kafka-python==2.0.2
influxdb-client==1.38.0
influxdb3-python==0.5.0
pandas==2.1.0
scikit-learn==1.3.0
fastapi==0.103.0
uvicorn==0.23.0
streamlit==1.27.0
plotly==5.17.0
python-dotenv==1.0.0
requests==2.31.0
tqdm==4.66.0
minio==7.2.9
pyarrow==15.0.0
polars==0.20.7
joblib>=1.3.0
numpy>=1.24.0
pydantic>=2.0.0
# ── Scheduler dependencies ────────────────────────────────────────────────────
APScheduler==3.10.4
docker>=6.1.0
